refactor: rename user and role column files for consistency;
update user deletion dialog to use correct property name
