+++ src/app/(admin)/roles/_components/RolesTable.tsx
@@ -4,7 +4,7 @@ import { useMemo } from "react";
 
 import { DataTable } from "@/shared/components/data-table/DataTable";
 import { useRoles } from "../_hooks/useRoles";
-import { columnsRoles } from "./column";
+import { columnsRoles } from "./roles.column";
 
 export default function RolesTable() {
   const { data: roles, isLoading, error } = useRoles();
